Rename: index to main
