+++ src/components/main/item/algorithmFilter.module.scss
