+++ src/components/main/item/sidebar.module.scss
